move general types to __init__
