+++ src/github_runner_manager/types_/__init__.py
@@ -2,3 +2,107 @@
 # See LICENSE file for licensing details.
 
 """Package containing modules with type definitions."""
+from typing import Optional
+
+from pydantic import BaseModel, MongoDsn, AnyHttpUrl, validator, IPvAnyAddress, Field
+
+
+class ReactiveConfig(BaseModel):
+    """Represents the configuration for reactive scheduling.
+
+    Attributes:
+        mq_uri: The URI of the MQ to use to spawn runners reactively.
+    """
+
+    mq_uri: MongoDsn
+
+
+class ProxyConfig(BaseModel):
+    """Proxy configuration.
+
+    Attributes:
+        aproxy_address: The address of aproxy snap instance if use_aproxy is enabled.
+        http: HTTP proxy address.
+        https: HTTPS proxy address.
+        no_proxy: Comma-separated list of hosts that should not be proxied.
+        use_aproxy: Whether aproxy should be used for the runners.
+    """
+
+    http: Optional[AnyHttpUrl]
+    https: Optional[AnyHttpUrl]
+    no_proxy: Optional[str]
+    use_aproxy: bool = False
+
+    @property
+    def aproxy_address(self) -> Optional[str]:
+        """Return the aproxy address."""
+        if self.use_aproxy:
+            proxy_address = self.http or self.https
+            # assert is only used to make mypy happy
+            assert (
+                proxy_address is not None and proxy_address.host is not None
+            )  # nosec for [B101:assert_used]
+            aproxy_address = (
+                proxy_address.host
+                if not proxy_address.port
+                else f"{proxy_address.host}:{proxy_address.port}"
+            )
+        else:
+            aproxy_address = None
+        return aproxy_address
+
+    @validator("use_aproxy")
+    @classmethod
+    def check_use_aproxy(cls, use_aproxy: bool, values: dict) -> bool:
+        """Validate the proxy configuration.
+
+        Args:
+            use_aproxy: Value of use_aproxy variable.
+            values: Values in the pydantic model.
+
+        Raises:
+            ValueError: if use_aproxy was set but no http/https was passed.
+
+        Returns:
+            Validated use_aproxy value.
+        """
+        if use_aproxy and not (values.get("http") or values.get("https")):
+            raise ValueError("aproxy requires http or https to be set")
+
+        return use_aproxy
+
+    def __bool__(self) -> bool:
+        """Return whether the proxy config is set.
+
+        Returns:
+            Whether the proxy config is set.
+        """
+        return bool(self.http or self.https)
+
+
+class SSHDebugConnection(BaseModel):
+    """SSH connection information for debug workflow.
+
+    Attributes:
+        host: The SSH relay server host IP address inside the VPN.
+        port: The SSH relay server port.
+        rsa_fingerprint: The host SSH server public RSA key fingerprint.
+        ed25519_fingerprint: The host SSH server public ed25519 key fingerprint.
+    """
+
+    host: IPvAnyAddress
+    port: int = Field(0, gt=0, le=65535)
+    rsa_fingerprint: str = Field(pattern="^SHA256:.*")
+    ed25519_fingerprint: str = Field(pattern="^SHA256:.*")
+
+
+class RepoPolicyComplianceConfig(BaseModel):
+    """Configuration for the repo policy compliance service.
+
+    Attributes:
+        token: Token for the repo policy compliance service.
+        url: URL of the repo policy compliance service.
+    """
+
+    token: str
+    url: AnyHttpUrl
